% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\docType{data}
\name{asdf_jags_out}
\alias{asdf_jags_out}
\title{Example data: asdf jags out}
\format{
An object of class \code{jagsUI} of length 24.
}
\usage{
asdf_jags_out
}
\description{
A simple model
}
\keyword{datasets}
