% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\docType{data}
\name{asdf_prior_jags_out}
\alias{asdf_prior_jags_out}
\title{Example data: asdf prior jags out}
\format{
An object of class \code{jagsUI} of length 24.
}
\usage{
asdf_prior_jags_out
}
\description{
A simple model, equivalent to that produced by the output produced by \verb{\link{skeleton}},
with the addition of prior samples for all parameters.
}
\keyword{datasets}
