% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{chaindens_df}
\alias{chaindens_df}
\title{By-chain kernel density of each column of a \code{data.frame}}
\usage{
chaindens_df(df, nline, parmfrow = NULL, ...)
}
\arguments{
\item{df}{Posterior \code{data.frame}}

\item{nline}{Number of chains}

\item{parmfrow}{Optional call to \code{par(mfrow)} for the number of rows & columns of plot window.  Returns the graphics device to previous state afterward.}

\item{...}{additional plotting arguments or arguments}
}
\value{
\code{NULL}
}
\description{
By-chain kernel density plot of each column of a posterior \code{data.frame}.
}
\examples{
out_df <- jags_df(asdf_jags_out)

b1 <- pull_post(out_df,"b1")
a <- pull_post(out_df,"a")

par(mfrow=c(3,1))
trace_df(a, nline=3)

chaindens_df(a, nline=3, parmfrow=c(3,1))
}
\seealso{
\link{tracedens_jags}, \link{trace_jags}, \link{trace_line}
}
\author{
Matt Tyers
}
