% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{chaindens_line}
\alias{chaindens_line}
\title{Simple by-chain kernel density plot}
\usage{
chaindens_line(x, nline, lwd = 1, main = "", ...)
}
\arguments{
\item{x}{Posterior vector}

\item{nline}{Number of chains}

\item{lwd}{Line width}

\item{main}{Plot title}

\item{...}{additional plotting arguments}
}
\value{
\code{NULL}
}
\description{
By-chain kernel density plot of a single parameter node.
}
\examples{
b1 <- jags_df(asdf_jags_out, p="b1")

chaindens_line(b1, nline=3, main="b1")
}
\seealso{
\link{tracedens_jags}, \link{chaindens_jags}, \link{chaindens_df}
}
\author{
Matt Tyers
}
