% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{cor_jags}
\alias{cor_jags}
\title{Correlation matrix from a JAGS object}
\usage{
cor_jags(x, p = NULL, exact = FALSE)
}
\arguments{
\item{x}{Output object returned from \code{jagsUI}}

\item{p}{Optional string to begin posterior names.  If \code{NULL} is used, all parameters will be used}

\item{exact}{Whether name must be an exact match (\code{TRUE}) or with initial sub-string matching only supplied characters (\code{FALSE}).
Defaults to \code{FALSE.}}
}
\value{
A 2-dimensional correlation matrix (n X n, where n is the number of parameter nodes)
}
\description{
Computes a correlation matrix of all MCMC samples from an object
returned by 'jagsUI', or an optional subset of parameter nodes.
}
\examples{
cor_jags(asdf_jags_out)
}
\seealso{
\link{plotcor_jags}
}
\author{
Matt Tyers
}
