% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{jags_df}
\alias{jags_df}
\title{Extract df}
\usage{
jags_df(x)
}
\arguments{
\item{x}{Output object from jagsUI::jags()}
}
\description{
Extract df from jagsUI output.
}
\examples{
out_df <- jags_df(asdf_jags_out)
}
\author{
Matt Tyers
}
