% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Logit log(x/(1-x)
}
\examples{
logit(0.5)
}
\seealso{
\link{expit}
}
\author{
Matt Tyers
}
