% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{nparam}
\alias{nparam}
\title{Number of parameters}
\usage{
nparam(x)
}
\arguments{
\item{x}{Output object from \code{jagsUI::jags()}}
}
\value{
A single numeric value giving the number of parameters.
}
\description{
Number of parameters saved in \code{jagsUI} output.
}
\examples{
head(jags_df(asdf_jags_out))

nparam(asdf_jags_out)
}
\seealso{
\link{nbyname}
}
\author{
Matt Tyers
}
