% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{plotRhats}
\alias{plotRhats}
\title{Plotting all Rhat values}
\usage{
plotRhats(
  x,
  p = NULL,
  n.eff = FALSE,
  fence = NULL,
  plotsequence = FALSE,
  splitarr = FALSE,
  margin = NULL,
  ...
)
}
\arguments{
\item{x}{Output object returned from \code{jagsUI}}

\item{p}{Optional vector of parameters to subset}

\item{n.eff}{Optionally, whether to plot \code{n.eff} instead of \code{Rhat}.  Defaults to \code{FALSE}.}

\item{fence}{Value of horizontal lines to overlay as reference.  Accepting the default value (\code{NULL}) will give \code{fence} values
of 1.1 (a commonly used value) and 1.01 for Rhat, or 100 and 500 for n.eff.}

\item{plotsequence}{Whether to plot parameter vectors (or matrices) in a sequence, running left to right, which may
be useful for time series models, etc.  If the default (\code{FALSE}) is used, a vertical cluster will be plotted
for each parameter, resulting in a simpler plot if there are many parameters.  Note that the Rhat values will still be
plotted in sequence if the default (\code{FALSE}) is used.}

\item{splitarr}{Whether to split 2+ dimensional parameter arrays by a given dimension, rather than plotting the full
array in one vertical cluster or continuous sequence.  This may be recommended in the case of large arrays.  Defaults to \code{FALSE}.}

\item{margin}{If splitarr is set to TRUE, which array margin to split by.  In the case of a 2-dimensional array, setting
\code{margin=2} will separate the array by column.  If the default (\code{NULL}) is accepted, the function will split by the smallest dimension,
therefore splitting into the fewest groups.}

\item{...}{additional plotting arguments}
}
\description{
Plotting all values of Rhat (or alternately n.eff) from an output object returned by \code{jagsUI}, or perhaps a subset of parameters.
This function is intended as a quick graphical check of which parameters have adequately converged.
}
\examples{
## plotting everything
plotRhats(SS_jags)
str(SS_jags$Rhat)  # the associated values

plotRhats(SS_jags, n.eff=TRUE)
str(SS_jags$n.eff)  # the associated values

## behavior of splitarr and margin are shown
plotRhats(SS2d_jags)
plotRhats(SS2d_jags, splitarr=TRUE, n.eff=TRUE)
plotRhats(SS2d_jags, splitarr=TRUE, margin=2)
str(SS2d_jags$Rhat) # the associated values

## plotsequence may be useful in the case of a sequence of values
plotRhats(SS2d_jags, p=c("trend", "cycle"), splitarr=TRUE, plotsequence=TRUE)
}
\seealso{
\link{traceworstRhat}, \link{check_Rhat}
}
\author{
Matt Tyers
}
