% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{pull_post}
\alias{pull_post}
\title{Subset from posterior df}
\usage{
pull_post(x, p)
}
\arguments{
\item{x}{Posterior df}

\item{p}{String to begin posterior names}
}
\description{
Extract a subset vector or df from posterior df, with parameter names starting with a given string.
}
\examples{
out_df <- jags_df(asdf_jags_out)

b1 <- pull_post(out_df,"b1")
a <- pull_post(out_df,"a")
}
\author{
Matt Tyers
}
