% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{qq_postpred}
\alias{qq_postpred}
\title{Quantile-quantile plot from posterior predictive distribution}
\usage{
qq_postpred(ypp, y, p = NULL, ...)
}
\arguments{
\item{ypp}{Either a matrix or \code{data.frame} of posterior samples, or a}

\item{...}{Optional plotting arguments}
}
\value{
\code{NULL}
}
\description{
Produces a quantile-quantile plot, calculated from the quantiles of
a vector of data with respect to the matrix of associated posterior
predictive distributions.

While not an omnibus posterior predictive check, this plot can be useful
for detecting an overparameterized model, or else improper specification
of observation error.  Like a traditional Q-Q plot, a well-specified model
will have points that lie close to the x=y line.  In the case of this
function, an overparametrized model will typically produce a plot with a
much shallower slope, with associated posterior predictive quantiles close
to 0.5.

It should be noted that this function will only produce meaningful results
with a vector of data, as opposed to a single value.

The posterior predictive distribution can be specified in two possible ways:
either a single output object from \code{jagsUI} with an associated parameter
name, or as a matrix or \code{data.frame} of posterior samples.
}
\examples{
# first, a quick look at the example data...
str(SS_data)
str(SS_out$sims.list$ypp)

# using a jagsUI object as ypp input
qq_postpred(ypp=SS_out, p="ypp", y=SS_data$y)

# using a matrix as ypp input
qq_postpred(ypp=SS_out$sims.list$ypp, y=SS_data$y)
}
\seealso{
\link{comparedens}, \link{comparecat}
}
\author{
Matt Tyers
}
