% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{rcolors}
\alias{rcolors}
\title{Random Colors}
\usage{
rcolors(n)
}
\arguments{
\item{n}{Vector length}
}
\value{
A vector of colors
}
\description{
Creates a vector of randomly-generated colors.
}
\examples{
n <- 1000
cols <- rcolors(n)
x <- runif(n)
y <- runif(n)
plot(x,y, col=cols, pch=16)
}
\author{
Matt Tyers
}
