% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\docType{data}
\name{SS2d_jags}
\alias{SS2d_jags}
\title{SS 2D jags out}
\format{
An object of class \code{jagsUI} of length 24.
}
\usage{
SS2d_jags
}
\description{
A more complicated time series model, with 2 dimensions
}
\keyword{datasets}
