% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\docType{data}
\name{SS_data}
\alias{SS_data}
\title{Example data: Time series associated with SS JAGS out}
\format{
An object of class \code{data.frame} with 41 rows and 2 columns.
}
\usage{
SS_data
}
\description{
The time series and time measurements associated with the time series model \verb{\link{SS_out}}.
}
\keyword{datasets}
