% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\docType{data}
\name{SS_out}
\alias{SS_out}
\title{Example data: SS JAGS out}
\format{
An object of class \code{jagsUI} of length 24.
}
\usage{
SS_out
}
\description{
A time series model with multiple observations of a single time series, a
nd with two stochastic cycle components.
}
\details{
This model is included partly to show a model with vectors or 2-dimensional
matrices of parameter nodes, and also to give an example of poor model convergence.
}
\keyword{datasets}
