% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{trace_line}
\alias{trace_line}
\title{Simple trace plot}
\usage{
trace_line(x, nline, lwd = 1, main = "", ...)
}
\arguments{
\item{x}{Posterior vector}

\item{nline}{Number of MCMC chains}

\item{lwd}{Line width}

\item{main}{Plot title}

\item{...}{additional plotting arguments}
}
\value{
\code{NULL}
}
\description{
Trace plot of a single parameter node.
}
\examples{
b1 <- jags_df(asdf_jags_out, p="b1")

trace_line(b1, nline=3, main="b1")
}
\seealso{
\link{tracedens_jags}, \link{trace_jags}, \link{trace_df}, \link{chaindens_line}
}
\author{
Matt Tyers
}
