% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{tracedens_jags}
\alias{tracedens_jags}
\title{Combination of trace plots and by-chain kernel densities of jagsUI object}
\usage{
tracedens_jags(x, p = NULL, parmfrow = NULL, lwd = 1, ...)
}
\arguments{
\item{x}{Posterior jagsUI object}

\item{p}{parameter name for subsetting: if this is specified, only parameters with names beginning with this string will be plotted.}

\item{parmfrow}{Optional call to par(mfrow) for the number of rows & columns of plot window.  Returns the graphics device to previous state afterward.}

\item{...}{additional plotting arguments}
}
\description{
Combination of trace plots and by-chain kernel densities of a whole jagsUI object, or subset.
}
\examples{

tracedens_jags(asdf_jags_out, parmfrow=c(4,2))
tracedens_jags(x=asdf_jags_out, p="a", parmfrow=c(3,1))
}
\author{
Matt Tyers
}
