% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\docType{data}
\name{ts_jags_out}
\alias{ts_jags_out}
\title{TS jags out}
\format{
An object of class \code{jagsUI} of length 24.
}
\usage{
ts_jags_out
}
\description{
A simple time series model
}
\keyword{datasets}
