% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\docType{data}
\name{xss}
\alias{xss}
\title{State-space X}
\format{
An object of class \code{numeric} of length 61.
}
\usage{
xss
}
\description{
The time variable associated with the time series models SS_jags and SS2d_jags
}
\keyword{datasets}
