% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{caterpillar}
\alias{caterpillar}
\title{Caterpillar plot}
\usage{
caterpillar(
  df,
  x = NA,
  median = T,
  ci = c(0.5, 0.95),
  lwd = 1,
  col = 4,
  add = F,
  xlab = "",
  ylab = "",
  main = "",
  xax = NA,
  ...
)
}
\arguments{
\item{df}{Posterior df}

\item{x}{Vector of X-coordinates for plotting.}

\item{median}{Whether to include medians}

\item{ci}{Vector of intervals to overlay.  Defaults to 50 percent and 95 percent.}

\item{lwd}{Base line width for plotting.  Defaults to 1.}

\item{col}{Color for plotting}

\item{add}{Whether to add to existing plot}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{main}{Plot title}

\item{xax}{vector of possible x-axis tick labels.  Defaults to the df column names}

\item{...}{additional plotting arguments or arguments to trace_line}
}
\description{
Caterpillar plot of posterior df.  This would be best suited to a single posterior df representing random effects or group means.
}
\note{
Calling this function with argument ci=0 (or F) is effectively a tricksy way of producing a plot with just a median line.
}
\examples{
out_df <- jags_df(asdf_jags_out)

b1 <- pull_post(out_df,"b1")
a <- pull_post(out_df,"a")

caterpillar(a)
caterpillar(a, ci=seq(.1,.9,by=.1))
caterpillar(a, lwd=2)
caterpillar(a, xax=c("effect 1", "effect 2", "effect 3"))
}
\author{
Matt Tyers
}
