% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{chaindens_jags}
\alias{chaindens_jags}
\title{By-chain kernel densities of \code{jagsUI} object}
\usage{
chaindens_jags(x, p = NULL, parmfrow = NULL, lwd = 1, ...)
}
\arguments{
\item{x}{Posterior \code{jagsUI} object}

\item{p}{parameter name for subsetting: if this is specified, only parameters with names beginning with this string will be plotted.}

\item{parmfrow}{Optional call to \code{par(mfrow)} for the number of rows & columns of plot window.  Returns the graphics device to previous state afterward.}

\item{lwd}{Line width for plotting.  Defaults to 1.}

\item{...}{additional plotting arguments}
}
\description{
By-chain kernel densities of a whole \code{jagsUI} object, or subset.
}
\examples{

chaindens_jags(asdf_jags_out, parmfrow=c(4,2))
chaindens_jags(x=asdf_jags_out, p="a", parmfrow=c(3,1))
}
\seealso{
\link{tracedens_jags}, \link{trace_jags}, \link{chaindens_line}, \link{chaindens_df}
}
\author{
Matt Tyers
}
