% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{check_Rhat}
\alias{check_Rhat}
\title{Quick summary of Rhat values by parameter name}
\usage{
check_Rhat(x, thresh = 1.1)
}
\arguments{
\item{x}{Output object from \code{jagsUI::jags()}}

\item{thresh}{Threshold value (defaults to 1.1)}
}
\value{
Numeric (named) giving the proportion of Rhat values below the given threshold.
}
\description{
Returns the mean number of Rhat values (by each parameter) that are less than a specified threshold criterion.
}
\examples{
check_Rhat(ts_jags_out)
}
\seealso{
\link{check_neff}, \link{traceworstRhat}, \link{plotRhats}
}
\author{
Matt Tyers
}
