% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{envelope}
\alias{envelope}
\title{Envelope plot}
\usage{
envelope(
  df,
  x = NA,
  median = T,
  ci = c(0.5, 0.95),
  col = 4,
  add = F,
  dark = 0.3,
  outline = F,
  xlab = "",
  ylab = "",
  main = "",
  ...
)
}
\arguments{
\item{df}{Posterior df}

\item{x}{Vector of X-coordinates for plotting.}

\item{median}{Whether to include median line}

\item{ci}{Vector of intervals to overlay.  Defaults to 50 percent and 95 percent.}

\item{add}{Whether to add to existing plot}

\item{dark}{Opacity (0-1) for envelopes.  Note that multiple overlapping intervals will darken the envelope.}

\item{outline}{Whether to just envelope outlines}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{main}{Plot title}

\item{...}{additional plotting arguments or arguments to trace_line}
}
\description{
Envelope plot of posterior df.  This would be best suited to a single posterior df representing a time series.  I've written and re-written this function many times!
}
\note{
Calling this function with argument ci=0 (or F) is effectively a tricksy way of producing a plot with just a median line.
}
\examples{
out_df <- jags_df(asdf_jags_out)

b1 <- pull_post(out_df,"b1")
a <- pull_post(out_df,"a")

envelope(a)
envelope(a, ci=seq(.1,.9,by=.1), dark=.15, lwd=3)
envelope(a, ci=seq(.1,.9,by=.1), dark=.2, outline=T)
envelope(a, ci=F, add=T, lwd=3)
}
\author{
Matt Tyers
}
