% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{expit}
\alias{expit}
\title{Expit, or inverse logit}
\usage{
expit(x)
}
\arguments{
\item{x}{A number}
}
\description{
Inverse logit, where logit is log(x/(1-x).  Expit is exp(x)/(1+exp(x))
}
\examples{
expit(0)
}
\author{
Matt Tyers
}
