% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{jags_df}
\alias{jags_df}
\title{Extract data.frame}
\usage{
jags_df(x)
}
\arguments{
\item{x}{Output object from \code{jagsUI::jags()}}
}
\value{
A \code{data.frame} with a column associated with each parameter and a row
associated with each MCMC iteration.
}
\description{
Extract \code{data.frame} from \code{jagsUI} output.
}
\examples{
out_df <- jags_df(asdf_jags_out)
}
\seealso{
\link{pull_post}
}
\author{
Matt Tyers
}
