% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{jags_df}
\alias{jags_df}
\title{Extract data.frame}
\usage{
jags_df(x)
}
\arguments{
\item{x}{Output object from \code{jagsUI::jags()}}
}
\value{
A \code{data.frame} with a column associated with each parameter and a row
associated with each MCMC iteration.
}
\description{
Extracts the posterior samples from \code{jagsUI} output in the form of
a \code{data.frame}.  This simpler construction has a few benefits: operations may
be more straightforward, and posterior objects will be smaller files and can be
written to an external table or .csv, etc.
}
\examples{
out_df <- jags_df(asdf_jags_out)
}
\seealso{
\link{pull_post}
}
\author{
Matt Tyers
}
