% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{jags_plist}
\alias{jags_plist}
\title{Plist}
\usage{
jags_plist(x, p = NULL)
}
\arguments{
\item{x}{jagsUI object}

\item{p}{String to subset parameter names, if a subset is desired}
}
\description{
Extract a list of nrep X nchain matrices, one for each parameter.
}
\examples{
out_plist <- jags_plist(asdf_jags_out)
str(out_plist)

a_plist <- jags_plist(asdf_jags_out, p="a")
str(a_plist)
}
\author{
Matt Tyers
}
