% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{nbyname}
\alias{nbyname}
\title{Number of parameters by parameter name}
\usage{
nbyname(x)
}
\arguments{
\item{x}{Output object from jagsUI::jags()}
}
\description{
Returns a summary of the numbers of parameters saved in jagsUI output by parameter name.
}
\examples{
head(jags_df(asdf_jags_out))

nbyname(asdf_jags_out)

nparam(ts_jags_out)
nbyname(ts_jags_out)
}
\author{
Matt Tyers
}
