% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{nparam}
\alias{nparam}
\title{Number of parameters}
\usage{
nparam(x)
}
\arguments{
\item{x}{Output object from jagsUI::jags()}
}
\description{
Number of parameters saved in jagsUI output.
}
\examples{
head(jags_df(asdf_jags_out))

nparam(asdf_jags_out)
}
\author{
Matt Tyers
}
