% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{skeleton}
\alias{skeleton}
\title{Skeleton}
\usage{
skeleton(NAME = "NAME")
}
\arguments{
\item{NAME}{Name to append to JAGS model object, etc.}
}
\description{
Provides a paste-able skeleton of code with an example JAGS model and output.
}
\examples{
skeleton("asdf")
}
\author{
Matt Tyers
}
