% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\docType{data}
\name{SS_jags}
\alias{SS_jags}
\title{SS jags out}
\format{
An object of class \code{jagsUI} of length 24.
}
\usage{
SS_jags
}
\description{
A more complicated time series model
}
\keyword{datasets}
