% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{trace_df}
\alias{trace_df}
\title{Traceplot of each column of a df}
\usage{
trace_df(df, nline, parmfrow = NULL, ...)
}
\arguments{
\item{df}{Posterior df}

\item{nline}{Number of chains}

\item{parmfrow}{Optional call to par(mfrow) for the number of rows & columns of plot window.  Returns the graphics device to previous state afterward.}

\item{...}{additional plotting arguments or arguments to trace_line}
}
\description{
Traceplot of each column of a posterior df.
}
\examples{
out_df <- jags_df(asdf_jags_out)

b1 <- pull_post(out_df,"b1")
a <- pull_post(out_df,"a")

par(mfrow=c(3,1))
trace_df(a, nline=3)

trace_df(a, nline=3, parmfrow=c(3,1))
}
\author{
Matt Tyers
}
