% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{trace_df}
\alias{trace_df}
\title{Traceplot of each column of a df}
\usage{
trace_df(df, nline, ...)
}
\arguments{
\item{df}{Posterior df}

\item{...}{additional plotting arguments or arguments to trace_line}

\item{p}{nline Number of chains}
}
\description{
Traceplot of each column of a posterior df.
}
\examples{
out_df <- jags_df(asdf_jags_out)

b1 <- pull_post(out_df,"b1")
a <- pull_post(out_df,"a")

trace_df(a, nline=3)
}
\author{
Matt Tyers
}
