% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{trace_jags}
\alias{trace_jags}
\title{Traceplot of jagsUI object}
\usage{
trace_jags(x, p = NULL, parmfrow = NULL, lwd = 1, ...)
}
\arguments{
\item{x}{Posterior \code{jagsUI} object}

\item{p}{parameter name for subsetting: if this is specified, only parameters with names beginning with this string will be plotted.}

\item{parmfrow}{Optional call to \code{par(mfrow)} for the number of rows & columns of plot window.  Returns the graphics device to previous state afterward.}

\item{lwd}{Line width for plotting.  Defaults to 1.}

\item{...}{additional plotting arguments}
}
\description{
Smarter traceplot of a whole \code{jagsUI} object, or subset.
}
\examples{

trace_jags(asdf_jags_out, parmfrow=c(4,2))
trace_jags(asdf_jags_out, p="a", parmfrow=c(3,1))
}
\seealso{
\link{tracedens_jags}, \link{pairstrace_jags}, \link{trace_df}, \link{trace_line}
}
\author{
Matt Tyers
}
