% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagshelper1.R
\name{trace_line}
\alias{trace_line}
\title{Simple traceplot}
\usage{
trace_line(x, nline, lwd = 1, main = "", ...)
}
\arguments{
\item{x}{Posterior vector}

\item{nline}{Number of chains}

\item{lwd}{Line width}

\item{main}{Plot title}

\item{...}{additional plotting arguments}
}
\description{
Traceplot of a single parameter.
}
\examples{
out_df <- jags_df(asdf_jags_out)

b1 <- pull_post(out_df,"b1")
a <- pull_post(out_df,"a")

trace_line(b1, nline=3, main="b1")
}
\seealso{
\link{tracedens_jags}, \link{trace_jags}, \link{trace_df}, \link{chaindens_line}
}
\author{
Matt Tyers
}
